Move some functions to a new module
